requirements.txt
black
flake8
isort
loguru
tqdm
pip
python-dotenv

psycopg2-binary
pandas

-e .
